feature: Allow reading partition key from JSON message body
Source connectors can only use the SQS message ID as a partition
key. This commit changes that by allowing the user to specify parsing
JSON message bodies for the purpose of extracting a key.

The user will be allowed to configure a type for the SQS message body.
They will also be allowed to configure a key for a JSON message body.
If the configured type is 'application/json' and a JSONPath value is
configured as the key then the message body will be parsed and an
attempt to extract the key using JSONPath syntax will be made.

Without additional configurations, the connector will fall back to
using the SQS message ID as the partition key.